Check commit

+++ examples/DIPDialect/correlation2D.cpp
@@ -65,12 +65,12 @@ bool testImplementation(int argc, char *argv[], std::ptrdiff_t x,
 
   // Define the kernel.
   float *kernelAlign = laplacianKernelAlign;
-  size_t kernelRows = laplacianKernelRows;
-  size_t kernelCols = laplacianKernelCols;
+  int kernelRows = laplacianKernelRows;
+  int kernelCols = laplacianKernelCols;
 
   // Define sizes and strides.
-  size_t sizesKernel[2] = {kernelRows, kernelCols};
-  size_t sizesOutput[2] = {static_cast<size_t>(image.rows), static_cast<size_t>(image.cols)};
+  intptr_t sizesKernel[2] = {kernelRows, kernelCols};
+  intptr_t sizesOutput[2] = {image.rows, image.cols};
 
   // Define memref containers.
   Img<float, 2> input(image);
